requirements.txt
flake8>=6.0.0
pytest>=7.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
